refactor: generic hook
